contracts: Loop 37 — B1 sealed green; open B2 (extract NegativeCacheStatistics, DTO-first)
B1 sealed: IWorkbenchSyncReadinessService CI-green on 84ff32d60 (Backend .NET
Tests 9m15s, Warning Gate /warnaserror, all gating jobs green). Workbench
tab-contract cluster complete.

Owner authorized B2 only. DTO-first step 1: extract NegativeCacheStatistics.
Gate 5/5 — self-contained BCL POCO (long/double/TimeSpan/DateTime/Dictionary;
no Core/EF), 3 consumers.

Extraction (class shared NegativeCachingService.cs with the service/options/iface,
so this is an extract-from-multi-type-file, not a whole-file move):
- Created Abstractions/DTOs/NegativeCacheStatistics.cs (ns
  TerraFusion.Abstractions.DTOs); left a breadcrumb comment in the source.
- InMemoryCacheStatisticsService (impl): add using Abstractions.DTOs (uses the
  type as a real symbol).
- ICacheStatisticsService (iface): swap Core.Services -> Abstractions.DTOs
  (sole use was this type).
- NegativeCachingService references the type only in comments -> NO using added
  (avoided an unused using under /warnaserror).

Trap noted: extract-from-multi-type-file requires checking each consumer
references the type as a real symbol before adding a using (comment-only refs
need none).

Build-safety (HR-4): dotnet unavailable locally — CI validates; claim limited to
"references migrated, structurally consistent." Lock remains PARTIALLY RELEASED
(R1 DTOs + R2 A-interfaces + R3/B1 + R4/B2 DTO); B3 + DEFERs + all else
ACTIVE-LOCKED.

Next: after green, B2 step 2 = ICacheStatisticsService. Updates CONTRACTS.md,
Loop 37 ledger, README.
