+++ backend/src/TerraFusion.Core/Interfaces/ICacheStatisticsService.cs
@@ -1,4 +1,4 @@
-using TerraFusion.Core.Services;
+using TerraFusion.Abstractions.DTOs;
 
 namespace TerraFusion.Core.Interfaces;
