+++ backend/src/TerraFusion.Core/Services/InMemoryCacheStatisticsService.cs
@@ -1,5 +1,6 @@
 using System.Collections.Concurrent;
 using TerraFusion.Core.Interfaces;
+using TerraFusion.Abstractions.DTOs;
 
 namespace TerraFusion.Core.Services;
