+++ backend/src/TerraFusion.Core/Services/NegativeCachingService.cs
@@ -417,22 +417,5 @@ internal class MissSentinelData
     public string RequestId { get; set; } = string.Empty;
 }
 
-/// <summary>
-/// Cache performance statistics
-/// </summary>
-public class NegativeCacheStatistics
-{
-    public long TotalRequests { get; set; }
-    public long CacheHits { get; set; }
-    public long CacheMisses { get; set; }
-    public long NegativeCacheHits { get; set; }
-    public long MissSentinelsSet { get; set; }
-    public long DatabaseQueriesPrevented { get; set; }
-    public double HitRatio => TotalRequests > 0 ? (double)CacheHits / TotalRequests * 100 : 0;
-    public double NegativeCacheEffectiveness => TotalRequests > 0 ? (double)NegativeCacheHits / TotalRequests * 100 : 0;
-    public TimeSpan AverageResponseTime { get; set; }
-    public DateTime StartTime { get; set; }
-    public DateTime LastReset { get; set; }
-    public Dictionary<string, long> ErrorCounts { get; set; } = new();
-    public Dictionary<string, double> ResponseTimesByOperation { get; set; } = new();
-}
+// NegativeCacheStatistics promoted to TerraFusion.Abstractions.DTOs
+// (Migrate Phase-1 / B2). See Abstractions/DTOs/NegativeCacheStatistics.cs.
